+++ CHANGELOG.md
@@ -5,7 +5,12 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.16.0] - 2023-07-25
+
+### Changed
+
+- Modified `FtpBruteForce`, `LdapBruteForce`, `RdpBruteForce` fields to align
+  with the event fields provided.
 
 ### Fixed
 
@@ -430,7 +435,7 @@ leading to a more streamlined system.
 
 - An initial version.
 
-[Unreleased]: https://github.com/petabi/review-database/compare/0.15.2...main
+[0.16.0]: https://github.com/petabi/review-database/compare/0.15.2...0.16.0
 [0.15.2]: https://github.com/petabi/review-database/compare/0.15.1...0.15.2
 [0.15.1]: https://github.com/petabi/review-database/compare/0.15.0...0.15.1
 [0.15.0]: https://github.com/petabi/review-database/compare/0.14.1...0.15.0